fix mix precision

+++ engine/embedding.py
@@ -7,7 +7,7 @@ from typing import Any
 class Embedding:
     def __init__(self, n:int, embed_dim:int) -> None:
         self.embed_dim = embed_dim
-        self.lookup_table = nx.uniform(low=-0.1, high=0.1, size=(n, self.embed_dim), dtype=nx.float32)
+        self.lookup_table = nx.uniform(low=-0.1, high=0.1, size=(n, self.embed_dim), dtype=nx.float16)
     
     def __eq__(self, value: object) -> bool:
         if not isinstance(value, Embedding):
